+++ examples/bootstrap.js
@@ -36,7 +36,7 @@ const MyControl = React.createClass({
       dialog = (
         <Dialog
           visible={this.state.visible}
-          animation="fade"
+          animation="slide-fade"
           maskAnimation="fade"
           onClose={this.onClose}
           style={{ width: 600 }}
@@ -61,9 +61,54 @@ const MyControl = React.createClass({
               </button>,
             ]
           }
-        >
-          <input />
-          <p>basic modal</p>
+        ><h4>Text in a modal</h4>
+          <p>Duis mollis, est non commodo luctus, nisi erat porttitor ligula.</p>
+          <hr/>
+          <h4>Overflowing text to show scroll behavior</h4>
+          <p>Cras mattis consectetur purus sit amet fermentum.
+            Cras justo odio, dapibus ac facilisis in,
+            egestas eget quam. Morbi leo risus, porta ac consectetur ac,
+            vestibulum at eros.
+          </p>
+          <p>Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Vivamus
+            sagittis lacus vel augue laoreet rutrum faucibus dolor auctor.
+          </p>
+          <p>Aenean lacinia bibendum nulla sed
+            consectetur. Praesent commodo cursus magna,
+            vel scelerisque nisl consectetur et. Donec sed odio dui.
+            Donec
+            ullamcorper nulla non metus auctor fringilla.
+          </p>
+          <p>Cras mattis consectetur purus sit amet fermentum.
+            Cras justo odio, dapibus ac facilisis in, egestas
+            eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros.
+          </p>
+          <p>Praesent commodo cursus
+            magna, vel scelerisque nisl consectetur et.
+            Vivamus sagittis lacus vel augue laoreet rutrum faucibus
+            dolor auctor.
+          </p>
+          <p>Aenean lacinia bibendum nulla sed consectetur.
+            Praesent commodo cursus magna, vel
+            scelerisque nisl consectetur et. Donec sed odio dui.
+            Donec ullamcorper nulla non metus auctor
+            fringilla.
+          </p>
+          <p>Cras mattis consectetur purus sit amet fermentum.
+            Cras justo odio, dapibus ac
+            facilisis in, egestas eget quam. Morbi leo risus,
+            porta ac consectetur ac, vestibulum at eros.
+          </p>
+          <p>
+            Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+            Vivamus sagittis lacus vel augue
+            laoreet rutrum faucibus dolor auctor.
+          </p>
+          <p>Aenean lacinia bibendum nulla sed consectetur. Praesent
+            commodo cursus magna, vel scelerisque nisl consectetur et.
+            Donec sed odio dui. Donec ullamcorper nulla
+            non metus auctor fringilla.
+          </p>
         </Dialog>
       );
     }
@@ -90,6 +135,7 @@ ReactDOM.render(
   <div>
     <h2>ant-design dialog</h2>
     <MyControl/>
-  </div>,
+  </div>
+  ,
   document.getElementById('__react-content')
 );